requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
feedparser==6.0.11
idna==3.8
pillow==10.4.0
reportlab==4.2.2
requests==2.32.3
sgmllib3k==1.0.0
soupsieve==2.6
urllib3==2.2.2
